+++ packages/builders/feather-icons-builder/svelte.config.js
@@ -10,7 +10,7 @@ const config = {
 	kit: {
 		adapter: adapter(),
 		package: {
-			exports: (file) => file === 'index.js'
+			// exports: (file) => file === 'index.js'
 			// files: (file) => !file.startsWith("internal"), //exclude internal Components
 		}
 		// hydrate the <div id="svelte"> element in src/app.html