+++ packages/builders/feather-icons-builder/CHANGELOG.md
@@ -1,1 +1,9 @@
 # @steeze-ui/feather-icons
+
+## 1.2.0
+
+### Minor Changes
+
+- improve typings
+- bundle sources in index.ts
+- shrinken package size